Duplicated currencies BUG solved
The PTAX form presents more than one currency id for the same currency.
For a few currencies (like MXN) the incorrect currency was being selected.
I implemented a workaround that gets the highest id.
This solution doesn't close this BUG, but for the great majority of currencies it works.

issue #19