Changes in API call to avoiding errors reported
The BCB API has changed and these changes generated errors in rbcb interface.

The API calls are simpler in order to handle future changes with less friction.

The $select attribute was removed, so all fields are downloaded and parsed into data.frame.

The indic argument now accept NULL so it is no longer necessary to specify indic to download data. If indic is NULL all available indics are downloaded.

#44
#45